@@ -6,7 +6,7 @@
   "main": "lib/main.js",
   "scripts": {
     "build": "tsc",
-    "lint": "eslint src/**/*.ts",
+    "lint": "eslint",
     "package": "npm-run-all package:*",
     "package:main": "ncc build -o dist --source-map --license licenses.txt",
     "package:deploy-path": "ncc build src/deploy-path.ts -o deploy-path/dist --source-map --license licenses.txt",
@@ -26,27 +26,29 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@actions/core": "^1.10.0",
+    "@actions/core": "^1.11.1",
     "@actions/exec": "^1.1.1",
-    "@actions/github": "^5.1.1",
-    "glob": "^8.1.0"
+    "@actions/github": "^6.0.0",
+    "glob": "^11.0.1"
   },
   "devDependencies": {
-    "@types/glob": "^8.0.1",
-    "@types/jest": "^27.5.0",
-    "@types/node": "^18.11.13",
-    "@typescript-eslint/eslint-plugin": "^5.46.0",
-    "@typescript-eslint/parser": "^5.46.0",
-    "@vercel/ncc": "^0.36.0",
-    "eslint": "^8.29.0",
-    "eslint-plugin-github": "^4.6.0",
-    "eslint-plugin-jest": "^27.1.6",
-    "jest": "^26.6.3",
-    "jest-circus": "^27.3.1",
+    "@eslint-community/eslint-plugin-eslint-comments": "^4.4.1",
+    "@eslint/js": "^9.22.0",
+    "@types/jest": "^29.5.14",
+    "@types/node": "^20.17.24",
+    "@typescript-eslint/eslint-plugin": "^8.26.1",
+    "@typescript-eslint/parser": "^8.26.1",
+    "@vercel/ncc": "^0.38.3",
+    "eslint": "^9.22.0",
+    "eslint-plugin-github": "^5.1.8",
+    "eslint-plugin-jest": "^28.11.0",
+    "jest": "^29.7.0",
+    "jest-circus": "^29.7.0",
     "js-yaml": "^4.1.0",
     "npm-run-all": "^4.1.5",
     "prettier": "3.5.2",
-    "ts-jest": "^26.5.6",
-    "typescript": "^4.9.4"
+    "ts-jest": "^29.2.6",
+    "typescript": "^5.8.2",
+    "typescript-eslint": "^8.26.1"
   }
 }
