@@ -1,31 +1,31 @@
-// default configuration tuned for development
-module.exports = {
-    parser: "@typescript-eslint/parser",
-    parserOptions: {
-      ecmaVersion: 9,
-      sourceType: "module",
-      project: "./tsconfig.json"
+import typescriptEslint, { configs as tsConfigs } from "typescript-eslint";
+import js from "@eslint/js";
+import comments from "@eslint-community/eslint-plugin-eslint-comments/configs";
+import tsParser from "@typescript-eslint/parser";
+
+export default typescriptEslint.config(
+  {
+    name: "ignores",
+    ignores: [
+      "deploy-path/dist/",
+      "dist/",
+      "lib/",
+      "node_models/",
+    ]
+  },
+  js.configs.recommended,
+  tsConfigs.recommended,
+  tsConfigs.stylistic,
+  comments.recommended,
+  {
+    name: "general rules",
+    languageOptions: {
+      parser: tsParser,
+      // We're currently using node v20, so 2023 looks good: https://node.green/#ES2023
+      ecmaVersion: 2023,
+      // TODO: It isn't clear if this is the right setting for this repo
+      sourceType: "module"
     },
-    plugins: ["@typescript-eslint", "jest" ],
-    env: {
-      node: true,
-      es6: true,
-      "jest/globals": true
-    },
-    settings: {
-      react: {
-        pragma: "React",
-        version: "detect"
-      }
-    },
-    ignorePatterns: [
-      "build/", "node_modules/"
-    ],
-    extends: [
-      "eslint:recommended",
-      "plugin:@typescript-eslint/recommended",
-      "plugin:eslint-comments/recommended"
-    ],
     rules: {
       "@typescript-eslint/explicit-module-boundary-types": "off",
       "@typescript-eslint/no-confusing-non-null-assertion": "error",
@@ -35,11 +35,11 @@ module.exports = {
       "@typescript-eslint/no-shadow": ["error", { builtinGlobals: false, hoist: "all", allow: [] }],
       "@typescript-eslint/no-unused-vars": ["warn", { args: "none", ignoreRestSiblings: true }],
       "@typescript-eslint/no-var-requires": "off",
-      curly: ["error", "multi-line", "consistent"],
+      "curly": ["error", "multi-line", "consistent"],
       "dot-notation": "error",
       "eol-last": "warn",
-      eqeqeq: ["error", "smart"],
-      "eslint-comments/no-unused-disable": "warn",
+      "eqeqeq": ["error", "smart"],
+      "@eslint-community/eslint-comments/no-unused-disable": "warn",
       "no-debugger": "off",
       "no-duplicate-imports": "error",
       "no-sequences": "error",
@@ -61,22 +61,8 @@ module.exports = {
       "prefer-rest-params": "error",
       "prefer-spread": "error",
       "quotes": [2, "double", { allowTemplateLiterals: true, avoidEscape: true }],
-      radix: "error",
-      semi: ["error", "always"]
-    },
-    overrides: [
-      { // some rules can be relaxed in tests
-        files: ["**/*.test.*"],
-        rules: {
-          "@typescript-eslint/no-non-null-assertion": "off",
-        }
-      },
-      {
-        files: ["**/setupTests.js", "**/webpack.config.js"],
-        rules: {
-          "@typescript-eslint/no-require-imports": "off",
-          "@typescript-eslint/no-var-requires": "off"
-        }
-      }
-    ]
-};
+      "radix": "error",
+      "semi": ["error", "always"]
+    }
+  },
+);
